requirements.txt
pandas==1.5.3
pykrx==1.0.39
psycopg2-binary==2.9.5
SQLAlchemy==1.4.44
python-dotenv==0.21.0
matplotlib==3.6.3